#3567 - Further splitting of components directory
