+++ ui/app/src/components/PublishingQueue.tsx
@@ -34,15 +34,15 @@ import {
   READY_FOR_LIVE,
   Selected
 } from "../models/publishing";
-import ConfirmDropdown from "./ConfirmDropdown";
+import ConfirmDropdown from "./UserControl/ConfirmDropdown";
 import FilterDropdown from "./FilterDropdown";
 import { setRequestForgeryToken } from "../utils/auth";
 import TablePagination from '@material-ui/core/TablePagination';
-import ErrorState from "./ErrorState";
-import EmptyState from "./EmptyState";
+import ErrorState from "./SystemStatus/ErrorState";
+import EmptyState from "./SystemStatus/EmptyState";
 import Typography from "@material-ui/core/Typography";
 import HighlightOffIcon from '@material-ui/icons/HighlightOff';
-import Spinner from "./Spinner";
+import Spinner from "./SystemStatus/Spinner";
 import RefreshIcon from '@material-ui/icons/Refresh';
 import Button from "@material-ui/core/Button";
