+++ ui/app/src/components/SystemStatus/ErrorState.tsx
@@ -21,7 +21,7 @@ import makeStyles from "@material-ui/core/styles/makeStyles";
 import Typography from "@material-ui/core/Typography";
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
 import Fab from "@material-ui/core/Fab";
-import crack from "../assets/full-crack.svg";
+import crack from "../../assets/full-crack.svg";
 import { defineMessages, useIntl } from "react-intl";
 import OpenInNewIcon from '@material-ui/icons/OpenInNew';
