+++ ui/app/src/components/SystemStatus/Gears.tsx
