+++ ui/app/src/components/SystemStatus/LoadingState.tsx
@@ -19,7 +19,7 @@ import React from "react";
 import makeStyles from "@material-ui/core/styles/makeStyles";
 import Typography from "@material-ui/core/Typography";
 import Gears from "./Gears";
-import { backgroundColor } from "../styles/theme";
+import { backgroundColor } from "../../styles/theme";
 
 const useStyles = makeStyles(() => ({
   loadingView: {
